Move extract_names() with exclude logic to Card
