+++ src/card.py
@@ -4,6 +4,7 @@ from typing import Union, Iterable, Optional, Tuple, NamedTuple, cast
 import re
 from pathlib import Path
 from itertools import chain
+from functools import reduce
 from dataclasses import dataclass
 from contextlib import contextmanager
 import yaml.parser
@@ -319,4 +320,21 @@ class Card:
                     yield Crew(credits, i)
                 case [*_]:
                     yield Match(cast(list[str|dict], row), i)
+
+def extract_names(matches: Iterable[Match]) -> set[Name]:
+    """
+    Return a set of all distinct participants given a list of Matches.
+    Match object may have an `x` key in their options. This must be a list of integer 1-based indices.
+    If present, the indices mark people to REMOVE from this participant list.
+    """
+    return reduce(lambda s1, s2: s1 | s2, (names_in_match(m) for m in matches))
+
+def names_in_match(mm: Match) -> set[Name]:
+    names: list[Optional[Name]] = list(mm.all_names())
+    exclude = set(mm.options.get('x', []))
+    return set(name
+               for i, name in enumerate(names)
+               if i + 1 not in exclude # exclude is 1-based
+               and name) # Otherwise the type is set[Name|None]
+
